requirements.txt
asgiref==3.7.2
Django==4.2.6
django-cors==0.1
django-environ==0.11.2
djangorestframework==3.14.0
Pillow==10.1.0
psycopg2==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
apparmor==3.1.6
appdirs==1.4.4
argcomplete==2.0.0
asgiref==3.7.2
attrs==22.2.0
autocommand==2.2.2
beautifulsoup4==4.12.2
build==0.10.0
CacheControl==0.13.1
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
coverage==6.5.0
cryptography==41.0.3
cssselect==1.2.0
dbus-python==1.3.2
distlib==0.3.7
Django==4.2.3
djangorestframework==3.14.0
docutils==0.20.1
fastjsonschema==2.18.0
filelock==3.12.2
idna==3.4
inflect==7.0.0
installer==0.7.0
jaraco.context==4.3.0
jaraco.functools==3.8.1
jaraco.text==3.11.1
LibAppArmor==3.1.6
lockfile==0.12.2
lxml==4.9.2
Mako==1.2.4
Markdown==3.4.4
MarkupSafe==2.1.3
meson==1.2.1
more-itertools==10.1.0
msgpack==1.0.5
numpy==1.25.1
ordered-set==4.1.0
packaging==23.1
Pillow==10.0.0
platformdirs==3.10.0
ply==3.11
pycairo==1.24.0
pycparser==2.21
pydantic==1.10.9
PyGObject==3.44.1
pyproject_hooks==1.0.0
pyserial==3.5
pytz==2023.3
pywal==3.4.0
Reflector==2023.6.28.0.36.1
requests==2.28.2
scour==0.38.2
sdkmanager==0.6.5
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.3
tomli==2.0.1
tqdm==4.66.1
trove-classifiers==2023.8.8
typing_extensions==4.7.1
urllib3==1.26.15
validate-pyproject==0.13.post1.dev0+gb752273.d20230520
virtualenv==20.24.3
zstandard==0.21.0
